Loan Attribute Assignment changes

+++ Entities/LoanUpdateRequestDto.cs
@@ -4,7 +4,7 @@ namespace kingsightapi.Entities
     {
         public long LoanKey { get; init; }
         public string LoanCode { get; init; } = string.Empty;
-        public int? LoanAliasKey { get; init; }
+        public long? LoanAliasKey { get; init; }
         public short? LoanRanking { get; init; }
         public string DummyLoanLink { get; init; } = string.Empty;
         public bool? IsLoanInterestApplicable { get; init; }
